# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "description": "회원가입",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "description": "접근 키, 시크릿 키",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ReqRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RespRegister"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.RespRegister"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.RespRegister"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/auth.RespRegister"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.ReqRegister": {
            "type": "object",
            "required": [
                "upbit_access_key",
                "upbit_secret_key"
            ],
            "properties": {
                "upbit_access_key": {
                    "type": "string"
                },
                "upbit_secret_key": {
                    "type": "string"
                }
            }
        },
        "auth.RespRegister": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}